Refactor Markdown page generation to be separate from paper.
This will make it easy to re-use this code to create the blog.
