+++ src/posts/paper/10-example-person-risk.ts
