+++ src/posts/paper/14-research-sources.ts
