+++ src/posts/paper/3-covid-transmission.ts
