+++ src/posts/paper/6-person-risk.ts
