+++ src/posts/paper/7-basic-method.ts
