+++ src/posts/paper/8-intermediate-method.ts
