+++ src/posts/paper/9-advanced-method.ts
