+++ src/posts/paper/img/ferretti-infectiousness.gif
