+++ src/posts/paper/img/transmission-dynamics.png
