+++ src/posts/post.ts
@@ -1,0 +1,9 @@
+// Common structs for posts (whitepaper, blog)
+
+export interface PostContent {
+  title: string
+  shortTitle?: string
+  content: string
+}
+
+export type PostMap = { [key: string]: PostContent }
